requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.30
alembic==1.13.1

python-dotenv==1.0.1

pydantic==2.7.1
pydantic-settings==2.2.1

google-generativeai==0.3.2

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

